split tests into new file; add additional checks
